fix: copy account address
