+++ src/app/pages/home/components/account-area.tsx
@@ -1,113 +1,12 @@
 import { memo } from 'react';
-import { FiCopy } from 'react-icons/fi';
 
-import { Box, Stack, StackProps, color, useClipboard } from '@stacks/ui';
-import { truncateMiddle } from '@stacks/ui-utils';
-import { UserAreaSelectors } from '@tests-legacy/integration/user-area.selectors';
+import { Stack, StackProps } from '@stacks/ui';
 
-import { useAnalytics } from '@app/common/hooks/analytics/use-analytics';
-import { Tooltip } from '@app/components/tooltip';
-import { Caption } from '@app/components/typography';
 import { CurrentAccountAvatar } from '@app/features/current-account/current-account-avatar';
 import { CurrentAccountName } from '@app/features/current-account/current-account-name';
-import { useCurrentAccountNamesQuery } from '@app/query/stacks/bns/bns.hooks';
-import { useCurrentAccountIndex } from '@app/store/accounts/account';
-import { useBtcAccountIndexAddressIndexZero } from '@app/store/accounts/blockchain/bitcoin/bitcoin-account.hooks';
 import { useCurrentAccount } from '@app/store/accounts/blockchain/stacks/stacks-account.hooks';
-import { useBitcoinFeature } from '@app/store/feature-flags/feature-flags.slice';
 
-const AccountBnsAddress = memo(() => {
-  const currentAccountNamesQuery = useCurrentAccountNamesQuery();
-
-  const bnsName = currentAccountNamesQuery.data?.names[0];
-  const { onCopy, hasCopied } = useClipboard(bnsName || '');
-  const analytics = useAnalytics();
-  const copyToClipboard = () => {
-    void analytics.track('copy_address_to_clipboard');
-    onCopy();
-  };
-
-  if (!bnsName) return null;
-
-  return (
-    <>
-      <Caption>{bnsName}</Caption>
-      <Tooltip
-        hideOnClick={false}
-        label={hasCopied ? 'Copied!' : 'Copy BNS name'}
-        placement="right"
-      >
-        <Stack>
-          <Box
-            _hover={{ cursor: 'pointer' }}
-            onClick={copyToClipboard}
-            size="12px"
-            color={color('text-caption')}
-            data-testid={UserAreaSelectors.AccountCopyBnsAddress}
-            as={FiCopy}
-          />
-        </Stack>
-      </Tooltip>
-    </>
-  );
-});
-
-const AccountAddress = memo((props: StackProps) => {
-  const currentAccount = useCurrentAccount();
-  const accountIndex = useCurrentAccountIndex();
-  const btcAddress = useBtcAccountIndexAddressIndexZero(accountIndex);
-  const { onCopy, hasCopied } = useClipboard(currentAccount?.address || '');
-  const analytics = useAnalytics();
-  const isBitcoinEnabled = useBitcoinFeature();
-
-  const copyToClipboard = () => {
-    void analytics.track('copy_address_to_clipboard');
-    onCopy();
-  };
-
-  return currentAccount ? (
-    <Stack isInline {...props}>
-      <Caption>{truncateMiddle(currentAccount.address, 4)}</Caption>
-      <Tooltip
-        hideOnClick={false}
-        label={hasCopied ? 'Copied!' : 'Copy Stacks address'}
-        placement="right"
-      >
-        <Stack>
-          <Box
-            _hover={{ cursor: 'pointer' }}
-            onClick={copyToClipboard}
-            size="12px"
-            color={color('text-caption')}
-            data-testid={UserAreaSelectors.AccountCopyAddress}
-            as={FiCopy}
-          />
-        </Stack>
-      </Tooltip>
-      {isBitcoinEnabled ? (
-        <>
-          <Caption>{truncateMiddle(btcAddress, 4)}</Caption>
-          <Tooltip
-            hideOnClick={false}
-            label={hasCopied ? 'Copied!' : 'Copy Bitcoin address'}
-            placement="right"
-          >
-            <Stack>
-              <Box
-                _hover={{ cursor: 'pointer' }}
-                onClick={copyToClipboard}
-                size="12px"
-                color={color('text-caption')}
-                data-testid={UserAreaSelectors.AccountCopyAddress}
-                as={FiCopy}
-              />
-            </Stack>
-          </Tooltip>
-        </>
-      ) : null}
-    </Stack>
-  ) : null;
-});
+import { AccountAddresses } from './account-addresses';
 
 export const CurrentAccount = memo((props: StackProps) => {
   const currentAccount = useCurrentAccount();
@@ -118,8 +17,7 @@ export const CurrentAccount = memo((props: StackProps) => {
       <Stack overflow="hidden" display="block" alignItems="flex-start" spacing="base-tight">
         <CurrentAccountName />
         <Stack isInline>
-          <AccountAddress />
-          <AccountBnsAddress />
+          <AccountAddresses />
         </Stack>
       </Stack>
     </Stack>
